Renaming areal to aerial
